feat(vs-code): change client and server code out of each package
