--- packages/lsp-sample/client/testFixture/diagnostics.txt
@@ -1,1 +1,0 @@
-ANY browsers, ANY OS.
