--- packages/lsp-sample/client/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "commonjs",
-		"target": "es2020",
-		"lib": ["es2020"],
-		"outDir": "out",
-		"rootDir": "src",
-		"sourceMap": true,
-	},
-	"include": ["src"],
-	"exclude": ["node_modules", ".vscode-test"]
-}
