--- packages/lsp-sample/server/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es2020",
-		"lib": ["es2020"],
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"sourceMap": true,
-		"strict": true,
-		"outDir": "out",
-		"rootDir": "src",
-	},
-	"include": ["src"],
-	"exclude": ["node_modules", ".vscode-test"]
-}
