+++ packages/lsp-sample/tsconfig.json
@@ -14,8 +14,4 @@
 		"node_modules",
 		".vscode-test"
 	],
-	"references": [
-		{ "path": "./client" },
-		{ "path": "./server" }
-	]
 }
